add a separate scraping module
